// src/types/DashboardStats.ts
export interface DashboardStats {
  activeSigningKeys: number;
  recentlyAddedKeys: number;
  totalUsers: number;
  recentlyAddedAdmins: number;
  recentlyAddedStaff: number;
  totalDocuments: number;
  recentlyAddedDocuments: number;
  signedDocuments: number;
  unsignedDocuments: number;
  pendingDocuments: number;
}